+++ src/helper/plugins/comms/get_blockchain_data.rs
@@ -114,7 +114,7 @@ pub fn api_receive_blockchain_server_tiles_by_height(
                                 height: block_data.h as u32,
                                 block_hash: hex_str_to_32_bytes(&block_data.x),
                                 block_time: block_data.t as u64,
-                                block_bits: block_data.b as u64,
+                                block_bits: block_data.b as u32,
                                 block_n_tx: block_data.n as u32,
                                 block_size: block_data.s as u32,
                                 block_fee: block_data.f as u64,
